requirements.txt
dj-static==0.0.6
Django==1.10
gunicorn==19.6.0
psycopg2==2.6.1
static3==0.7.0
django-ckeditor
django-localflavor==1.3
pydrive
boto
oauth2client